+++ task/controller/pkg/gitrestclient/export_test.go
@@ -1,0 +1,16 @@
+// Copyright (c) 2026 Benjamin Borbe All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package gitrestclient
+
+import "time"
+
+// NewGitRestClientForTest creates a GitRestClient with a custom backoff for use in tests.
+// Pass a function returning 0 or 1ms to make retry tests run fast.
+func NewGitRestClientForTest(
+	baseURL string,
+	backoff func(attempt int) time.Duration,
+) GitRestClient {
+	return newGitRestClientWithBackoff(baseURL, backoff)
+}